starlight/wallet: add loading state to config form
once the submit btn is clicked, we set `showSpinner: true`, which causes
the submit btn to become disabled and changes it's content to 'Configuring'
with a spinning icon.

if the server responds with an error, the btn will change into the error
state. if it is successful, a new page will render, so we don't need to
worry about un-setting the loading state.

Closes #5920

Author: jordynchuhaloff <jchuhaloff@gmail.com>
Date: Tue Oct 16 12:25:55 2018 -0700
upstream:2a6fb39577c45e7465675ccc2a1f5b5577a08265