starlight/wallet: move isHost filed to ChannelOp
Right now there's a bug where if a channel is closed and then reopened with
the
other party as Host (which is totally possible), getChannelActivity
incorrectly
sets "isHost" on the entire history activity to be the same. The Op itself
should keep track of whether the party was the Host at the time the op was
created.

Closes #6021

Author: jordynchuhaloff <jchuhaloff@gmail.com>
Date: Thu Oct 18 13:18:31 2018 -0700
upstream:fb50b1e4e8741d5ead2778dd1a5d4575af061d69