starlight/doc: add message authentication to spec
This adds a `MessageSignature` field to every message, which is a signature
from the sender's channel-specific public key on the body of the message
(i.e., all of the message-specific fields, including the `ChannelID`). This
prevents messages from being forged or malleated by a man-in-the-middle
attacker.

Closes #6016

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Sat Oct 20 13:16:02 2018 -0700
upstream:4d836e9a82a5756c03c7925e781ec8b1d90b3cf7