starlight: return error when channel exists
This causes the agent to return a 500 error when a user attempts to create
a channel that already exists. Previously, it returned a 205 error, which
confused the client (since it interprets 20x codes as success).
Distinguishing between these cases using the HTTP code probably is not
necessary at this point anyway.

Closes #6099

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Tue Oct 23 08:06:50 2018 -0700
upstream:cdc6430ada465657676b1f4c985304f008dcbd0b