starlight/wallet: look up account when sending to address
This changes the send payment flow to allow sending to addresses, not just
accounts. The wallet calls the `api/find-account` endpoint to look up the
Stellar account for that address.

This is probably not the best possible implementation of this feature (for
one thing, ideally we would get address lookup at the time the address is
typed in, rather than when the user clicks Send, so we can validate it for
the user), but it's the simplest, and unblocks the demo flow for now.

Blocked on #6081.

Closes #6082

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Tue Oct 23 11:52:19 2018 -0700
upstream:66012d5ba890a93081f9e9f051a0f4d26db22bd5