starlight/wallet: fix bug when reopening closed channel
This fixes a bug where, if a channel was closed, reopened, and then closed
again, the incoming money from the withdrawal was not recognized as a
channel withdrawal (because the new escrow account ID was not tracked in
the data store). It also forecloses an even edgier case in which a Host
closes a channel, recreates the escrow account, and then sends money to the
Guest (which under the old code would look like a channel withdrawal to the
Guest, and cause unexpected behavior).

Closes #5793

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Tue Oct 23 15:10:20 2018 -0700
upstream:617fd357a258d8d82eca80da37819915411ca6bf