starlight: check merge operation source for account closing
Fixes a bug where any txs that had the wallet account as the source with
merge operations would cause the agent to reset. Instead, we only want to
reset the agent when the wallet account itself is the source of the merge
operation.

Closes #6153

Author: Vicki Niu <vicki.niu@gmail.com>
Date: Tue Oct 23 20:43:46 2018 -0700
upstream:ea20e135760f65141f4cff76161b8abdf2440b8b