starlight/wallet: properly handle demo vs custom server on init
Previously, we were defaulting the server to our demo url in a bunch of
places.
This meant that you could click 'Provide server URL', leave it blank, and
submit. It failed silently by setting your server to the demo server.

Now, we will rely on the boolean `DemoServer` to let the `config.init` side
effect know whether to use the param `HorizonURL` or to overwrite that with
the
demo url.

I also made the Username and Password fields required in the form.

Closes #6173

Author: jordynchuhaloff <jchuhaloff@gmail.com>
Date: Wed Oct 24 13:20:05 2018 -0700
upstream:bb9ef8884daf476d27e628c865ecb757bc62a11c