starlight/wallet: use big.js to show stroops
It turns out that numeral.js's format returns `NaN` for small numbers.
This affects any number with 6 or more leading 0s, which incidentally
appears to be when the chrome console starts displaying numbers in
exponential form. Since a stroop is `0.0000001`, we need to handle
these small numbers. This PR adds Big.js in order to render them as a
string in non-exponent form.

https://github.com/adamwdraper/Numeral-js/issues/596#issuecomment-407763186

Closes #6187

Author: jessicard <jessicard@mac.com>
Date: Wed Oct 24 16:26:43 2018 -0700
upstream:7bfb0d56a0492035c60d35411fa94d93223528fb