starlight/wallet: update wallet on withdrawal
Previously, the event handler did not dispatch any action when it got an
update saying that a withdrawal had been added to the wallet account.  This
was intended to avoid adding a redundant activity to the wallet view, since
the withdrawal activity would be included in the channel activity. However,
this meant that the wallet's _balance_ was not being updated upon the
withdrawal.

This fixes that by dispatching an action that simply updates the balance
when there is a withdrawal.

Closes #6202

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Wed Oct 24 20:28:24 2018 -0700
upstream:b9fa2879a637bc8443f84062b9236f63032cd4e6