starlight: fix settleonlywithhost bugs
This fixes a bug where, if the Guest sent a channel payment that caused its
balance to be exhausted, it caused a signature verification error when the
Guest attempted to handle the PaymentAcceptMsg. It also fixes a bug where
the channel's original FundingTime was always used as the time for the
SettleOnlyWithHostTx, even when it was used as part of a payment.

Closes #6205

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Wed Oct 24 23:19:10 2018 -0700
upstream:e1aecdb89549966b1d61d79142c8e29889f8eda9