starlight/wallet: add ability to version our Redux store
If we make any breaking changes to our Redux store, we need to increment
REDUX_STORE_VERSION.

On page reload, this new migrate function will run. It checks the version
of
the client's Redux store, and if that does not match REDUX_STORE_VERSION,
it
resets all slices of the store, with the exception of config, lifecycle,
and
_persist. (If there was no change, it returns the existing state.)

Since we're keeping the config and lifecycle slices, the user won't be
logged
out. All of the events will be refreshed, though. (If we ever make changes
to
the config or lifecycle slices, we will have to remove them from here for
that
version change.)

Closes #6230

Author: jordynchuhaloff <jchuhaloff@gmail.com>
Date: Thu Oct 25 17:23:05 2018 -0700
upstream:b3ce6eb9f315278d9e111dcff9316d8160b4ae55