starlight: only stream txs for wallet account once tx has been submitted
The vast majority of our requests to the Horizon streaming endpoints come
from retrying requests to stream txs for an account that has not yet been
created. This change adds a channel that the `watchWalletAcct` goroutine
blocks on so as to not request a stream until the account-creating tx has
at least been successfully submitted.

Closes #6278

Author: Vicki Niu <vicki.niu@gmail.com>
Date: Tue Oct 30 12:48:27 2018 -0700
upstream:9809dd79a9cd727c8e5de7986023152ffaf1169f