starlight: resolve channel creation conflict
This change resolves conflict when two parties try to propose a channel
with each other simultaneously. The "winning" party is the one with the
earlier `FundingTime`: the other party cleans up their channel and accepts
the incoming `ChannelProposeMsg`.

This is waiting on some changes to error formatting and handling in
https://github.com/interstellar/i10r/pull/6224

Closes #6197

Author: Vicki Niu <vicki.niu@gmail.com>
Date: Thu Nov 1 09:31:07 2018 -0700
upstream:dfacef6a68b2cfffcf898b6ac5f4f109f8d20d41