starlight/wallet: make sdk an npm package
This turns the `client` folder into its own `starlight-sdk` npm package. It
also makes some changes to the client to allow it to work in Node as well
as in the browser, including adding cookie management to it.

Closes #6376

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Wed Nov 7 12:56:26 2018 -0800
upstream:ec1bb81ab58ca1976c2d59f60dcf401583e5f59d