starlight/doc: avoid need for Host to have a server
This changes the protocol to eliminate the need for Host to have a server,
by having Host receive messages by sending long-polling `GET` requests to
Guest. It also makes some changes to how messages are signed, to ensure
that Host is properly authenticated.

This is the first numbered update to the protocol—it changes the protocol
to version 2.

Closes #6388

Author: Daniel Robinson <danrobinson010@gmail.com>
Date: Mon Nov 19 12:31:23 2018 -0800
upstream:365d333fe0743a8edd7d56d741d4c0461f04720e